requirements.txt
setuptools
numpy
vtkbone
bonelab
yaml
pyyaml
pytorch